Rename folder
`modal-component` -> `modal`
