+++ solidity-v1/dashboard/src/components/modal/Modal.jsx
