+++ solidity-v1/dashboard/src/components/modal/coverage-pools/components/index.js
