+++ solidity-v1/dashboard/src/components/modal/coverage-pools/deposit/WarningBeforeDeposit.jsx
