+++ solidity-v1/dashboard/src/components/modal/index.js
