+++ solidity-v1/dashboard/src/components/modal/staking/CopyStake.jsx
