+++ solidity-v1/dashboard/src/components/modal/staking/index.js
