+++ solidity-v1/dashboard/src/components/modal/wallets/SelectedWalletModal.jsx
