+++ solidity-v1/dashboard/src/components/modal/wallets/index.js
